refactor(flags): rename repository files to PascalCase convention
